Masterclass

+++ src/GUI/MasterclassBeheer.java
@@ -64,12 +64,12 @@ public class MasterclassBeheer extends javax.swing.JFrame {
             while (res.next()) {
                  int M_Code = res.getInt("M_Code");
                  int Niveau = res.getInt("Niveau");
-                 int Prijs = res.getInt("Prijs");
+                 Double Prijs = res.getDouble("Prijs");
                  Date Datum = res.getDate("Datum");
                  int Rating = res.getInt("Minimale_rating");
                  int Docent = res.getInt("Docent");
                  int Locatie = res.getInt("Vindt_plaats_in");
-                 masterclass = new MasterclassZoeken(res.getInt("M_Code"), res.getInt("Niveau"), res.getInt("Prijs"), res.getDate("Datum"), res.getInt("Minimale_rating"), res.getInt("Docent"), res.getInt("Vindt_plaats_in"));
+                 masterclass = new MasterclassZoeken(res.getInt("M_Code"), res.getInt("Niveau"), res.getDouble("Prijs"), res.getDate("Datum"), res.getInt("Minimale_rating"), res.getInt("Docent"), res.getInt("Vindt_plaats_in"));
                  masterclasses.add(masterclass);
                  String DateFormat = DateUtil.fromSqlDateToString(Datum);
                  String[] mczoek = new String[]{"" + M_Code, "" + Niveau, "" + Prijs, "" + DateFormat, "" + Rating, "" + Docent, "" + Locatie};
@@ -211,7 +211,7 @@ public class MasterclassBeheer extends javax.swing.JFrame {
                     //Maakt een nieuwe deelnemer met alle bijbehorende attributen.
                     masterclass = new MasterclassZoeken(res.getInt("M_Code"),
                             res.getInt("Niveau"),
-                            res.getInt("Prijs"),
+                            res.getDouble("Prijs"),
                             res.getDate("Datum"),
                             res.getInt("Minimale_rating"),
                             res.getInt("Docent"),
